requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-writable-nested==0.7.0
django-filter==23.3
